:art: Improve structure

+++ src/main/java/com/example/xrefintegration/repository/ArticleRepository.java
@@ -1,10 +1,10 @@
 package com.example.xrefintegration.repository;
 
-import com.example.xrefintegration.model.Article;
+import com.example.xrefintegration.model.article.Article;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 
 @Repository
-public interface ArticleRepository extends JpaRepository<Article, Integer> {
+public interface ArticleRepository extends JpaRepository<Article, Long> {
 }
